Use Stream or LazyList depending on Scala version
