+++ core/shared/src/test/scala/fs2/ChunkGenerators.scala
@@ -23,11 +23,10 @@ package fs2
 
 import cats.data.Chain
 import scala.reflect.ClassTag
-import scala.collection.immutable.{Stream => StdStream}
 import scodec.bits.ByteVector
 import java.nio.{Buffer => JBuffer, CharBuffer => JCharBuffer, ByteBuffer => JByteBuffer}
 
-import org.scalacheck.{Arbitrary, Cogen, Gen, Shrink}
+import org.scalacheck.{Arbitrary, Cogen, Gen}
 import Arbitrary.arbitrary
 
 trait ChunkGeneratorsLowPriority1 extends MiscellaneousGenerators {
@@ -74,7 +73,7 @@ trait ChunkGeneratorsLowPriority extends ChunkGeneratorsLowPriority1 {
     )
 }
 
-trait ChunkGenerators extends ChunkGeneratorsLowPriority {
+trait ChunkGenerators extends ChunkGeneratorsLowPriority with ChunkGeneratorsCompat {
   private def arrayChunkGenerator[A](genA: Gen[A])(implicit ct: ClassTag[A]): Gen[Chunk[A]] =
     for {
       values <- Gen.listOf(genA).map(_.toArray)
@@ -149,40 +148,6 @@ trait ChunkGenerators extends ChunkGeneratorsLowPriority {
 
   implicit def cogenChunk[A: Cogen]: Cogen[Chunk[A]] =
     Cogen[List[A]].contramap(_.toList)
-
-  implicit def shrinkChunk[A]: Shrink[Chunk[A]] =
-    Shrink[Chunk[A]](c => removeChunks(c.size, c))
-
-  // The removeChunks function and the interleave function were ported from Scalacheck,
-  // from Shrink.scala, licensed under a Revised BSD license.
-  //
-  // /*-------------------------------------------------------------------------*\
-  // **  ScalaCheck                                                             **
-  // **  Copyright (c) 2007-2016 Rickard Nilsson. All rights reserved.          **
-  // **  http://www.scalacheck.org                                              **
-  // **                                                                         **
-  // **  This software is released under the terms of the Revised BSD License.  **
-  // **  There is NO WARRANTY. See the file LICENSE for the full text.          **
-  // \*------------------------------------------------------------------------ */
-
-  private def removeChunks[A](size: Int, xs: Chunk[A]): StdStream[Chunk[A]] =
-    if (xs.isEmpty) StdStream.empty
-    else if (xs.size == 1) StdStream(Chunk.empty)
-    else {
-      val n1 = size / 2
-      val n2 = size - n1
-      lazy val xs1 = xs.take(n1)
-      lazy val xs2 = xs.drop(n1)
-      lazy val xs3 =
-        for (ys1 <- removeChunks(n1, xs1) if !ys1.isEmpty) yield Chunk.Queue(ys1, xs2)
-      lazy val xs4 =
-        for (ys2 <- removeChunks(n2, xs2) if !ys2.isEmpty) yield Chunk.Queue(xs1, ys2)
-
-      StdStream.cons(xs1, StdStream.cons(xs2, interleave(xs3, xs4)))
-    }
-
-  private def interleave[A](xs: StdStream[A], ys: StdStream[A]): StdStream[A] =
-    if (xs.isEmpty) ys else if (ys.isEmpty) xs else xs.head +: ys.head +: (xs.tail ++ ys.tail)
 }
 
 object ChunkGenerators extends ChunkGenerators
